+++ src/main/java/com/hunteryavitz/blockchainapi/entities/healthmetric/HealthMetricProductionModel.java
@@ -2,36 +2,68 @@ package com.hunteryavitz.blockchainapi.entities.healthmetric;
 
 import lombok.Data;
 
+/**
+ * The HealthMetricProductionModel class is a model for the production health metric.
+ */
 @Data
 public class HealthMetricProductionModel {
 
+    /**
+     * The data for the production health metric.
+     */
     private Data data;
 
+    /**
+     * The constructor for the HealthMetricProductionModel class.
+     */
     public HealthMetricProductionModel() {
         data = new Data();
     }
 
+    /**
+     * Sets the data for the production health metric.
+     * @param data the data for the production health metric
+     */
     public void setDataset1(int[] data) {
         this.data.setData1(data);
     }
 
+    /**
+     * Sets the data for the production health metric.
+     * @param data the data for the production health metric
+     */
     public void setDataset2(int[] data) {
         this.data.setData2(data);
     }
 
-    class Data {
+    /**
+     * The Data class is a model for the data of the production health metric.
+     */
+    static class Data {
         private int[] dataset1;
         private int[] dataset2;
 
+        /**
+         * Sets the data for the production health metric.
+         * @param data the data for the production health metric
+         * */
         public void setData1(int[] data) {
             this.dataset1 = data;
         }
 
+        /**
+         * Sets the data for the production health metric.
+         * @param data the data for the production health metric
+         */
         public void setData2(int[] data) {
             this.dataset2 = data;
         }
     }
 
+    /**
+     * Returns the data as a JSON string.
+     * @return the data as a JSON string
+     */
     public String asJson() {
         return "{" +
                 "\"data\": {" +
@@ -41,6 +73,11 @@ public class HealthMetricProductionModel {
                 "}";
     }
 
+    /**
+     * Returns the data as a JSON array.
+     * @param dataset1 the data to be converted to a JSON array
+     * @return the data as a JSON array
+     */
     private String asJsonArray(int[] dataset1) {
         StringBuilder sb = new StringBuilder();
         sb.append("[");