+++ src/main/java/com/hunteryavitz/blockchainapi/utils/HealthMetricUtils.java
@@ -1,5 +1,8 @@
 package com.hunteryavitz.blockchainapi.utils;
 
+/**
+ * The HealthMetricUtils class is a utility class for the health metric.
+ */
 public class HealthMetricUtils {
 
 }