+++ src/main/java/com/hunteryavitz/blockchainapi/utils/Utils.java
@@ -132,8 +132,12 @@ public class Utils {
         return updatedLiveness;
     }
 
+    /**
+     * The getContaminationSpectrum method returns the contamination spectrum of a node.
+     * @param contaminationSpectrum The contamination spectrum of the node.
+     * @return The contamination spectrum of the node.
+     */
     public static Integer[] getCopntaminationSpectrum(Integer[] contaminationSpectrum) {
         return null;
-
     }
 }