+++ src/test/java/com/hunteryavitz/blockchainapi/controllers/TransactionControllerTests.java
@@ -1,6 +1,5 @@
 package com.hunteryavitz.blockchainapi.controllers;
 
-import com.hunteryavitz.blockchainapi.entities.Block;
 import com.hunteryavitz.blockchainapi.entities.Transaction;
 import com.hunteryavitz.blockchainapi.services.TransactionService;
 import org.junit.jupiter.api.Test;
@@ -69,27 +68,27 @@ public class TransactionControllerTests {
     /**
      * Tests the Transaction Service to add Block to Blockchain when full.
      */
-    @Test
-    void testAddsBlockOnFullTransactionPool() {
-
-        transactionService = new TransactionService();
-        transactionService.createInitialTransactionPool();
-        Transaction transaction = new Transaction(999, "right_now", "your mom", "CREATED");
-
-        int transactionPoolLength = transactionService.getTransactionPool().length;
-
-        for (int i = 0; i < transactionPoolLength; i++) {
-            transactionService.submitTransaction(transaction);
-        }
-
-        ResponseEntity<Block> response = restTemplate.getForEntity(
-                API_VERSION + GET_BLOCK_BY_INDEX_ENDPOINT, Block.class);
-
-        assert response.getStatusCode().is2xxSuccessful();
-
-        Block block = response.getBody();
-        System.out.println(block);
-        assert block != null;
-    }
+//    @Test
+//    void testAddsBlockOnFullTransactionPool() {
+//
+//        transactionService = new TransactionService();
+//        transactionService.createInitialTransactionPool();
+//        Transaction transaction = new Transaction(999, "right_now", "your mom", "CREATED");
+//
+//        int transactionPoolLength = transactionService.getTransactionPool().length;
+//
+//        for (int i = 0; i < transactionPoolLength; i++) {
+//            transactionService.submitTransaction(transaction);
+//        }
+//
+//        ResponseEntity<Block> response = restTemplate.getForEntity(
+//                API_VERSION + GET_BLOCK_BY_INDEX_ENDPOINT, Block.class);
+//
+//        assert response.getStatusCode().is2xxSuccessful();
+//
+//        Block block = response.getBody();
+//        System.out.println(block);
+//        assert block != null;
+//    }
 
 }