+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Blockchain API
 Unreleased
 
-*v0.0.14*
+*v0.0.15*
 
-**10/19/23**
+**10/20/23**
 
 - [ADD Endpoint for Health Check](#47)
 - [ADD Endpoint for Production Check](#48)